refactor: simplify Game model
