+++ src/models/game.ts
@@ -1,0 +1,11 @@
+import { ObjectId } from "mongodb";
+
+export default class Game {
+    constructor(
+        public name: string,
+        public price: number,
+        public category: string,
+        public id?: ObjectId
+    ) {
+    }
+}
